fixed query params findMostKudosedSkill & findMostKudosedProof

+++ src/main/java/com/teamolha/talantino/skill/model/request/FullSkillDTO.java
@@ -7,6 +7,6 @@ public record FullSkillDTO(
         Long id,
         String label,
         String icon,
-        Integer totalKudos
+        Long totalKudos
 ) {
 }
